+++ CHANGELOG.md
@@ -12,6 +12,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Push attendance when a student is on stage
 
+### Changed
+
+- Use custom policy to generate cloudfront signed url
+
 ## [4.0.0-beta.4] - 2022-05-17
 
 ### Added